requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.22
langchainhub==0.1.15
faiss-cpu==1.12.0
tiktoken==0.7.0
openai>=1.40.0
youtube-transcript-api==1.2.2
google-api-python-client==2.149.0
python-dotenv==1.0.1
fastapi>=0.95
uvicorn[standard]>=0.22
streamlit==1.37.0
requests==2.32.3

